requirements.txt
pytest==6.2.5
requests==2.26.0
allpairspy==2.5.0
selenium==4.1.0
allure-pytest==2.9.45